+++ index.ts
@@ -224,3 +224,4 @@ export import emaildataplane = require("oci-emaildataplane");
 export import clusterplacementgroups = require("oci-clusterplacementgroups");
 export import marketplaceprivateoffer = require("oci-marketplaceprivateoffer");
 export import resourcescheduler = require("oci-resourcescheduler");
+export import demandsignal = require("oci-demandsignal");