+++ lib/capacitymanagement/lib/model/sort-order.ts
@@ -19,7 +19,13 @@ import common = require("oci-common");
  **/
 export enum SortOrder {
   Asc = "ASC",
-  Desc = "DESC"
+  Desc = "DESC",
+
+  /**
+   * This value is used if a service returns a value for this enum that is not recognized by this
+   * version of the SDK.
+   */
+  UnknownValue = "UNKNOWN_VALUE"
 }
 
 export namespace SortOrder {