+++ lib/databasemanagement/lib/model/database-password-credential-details.ts
@@ -41,6 +41,7 @@ export namespace DatabasePasswordCredentialDetails {
   export enum Role {
     Normal = "NORMAL",
     Sysdba = "SYSDBA",
+    Sysdg = "SYSDG",
     /**
      * This value is used if a service returns a value for this enum that is not recognized by this
      * version of the SDK.