+++ lib/databasemanagement/lib/model/database-ssl-connection-credentials.ts
@@ -60,6 +60,7 @@ export namespace DatabaseSslConnectionCredentials {
   export enum Role {
     Sysdba = "SYSDBA",
     Normal = "NORMAL",
+    Sysdg = "SYSDG",
     /**
      * This value is used if a service returns a value for this enum that is not recognized by this
      * version of the SDK.