+++ lib/jmsjavadownloads/lib/model/create-java-download-report-details.ts
@@ -1,6 +1,6 @@
 /**
  * Java Management Service Download API
- * The APIs for the download engine of the Java Management Service.
+ * The APIs for the <a href="https://docs.oracle.com/en-us/iaas/jms/doc/java-download.html">Java Download</a> feature of Java Management Service.
  * OpenAPI spec version: 20230601
  *
  *
@@ -35,17 +35,29 @@ export interface CreateJavaDownloadReportDetails {
    */
   "timeEnd"?: Date;
   /**
-   * The property to be used for sorting the reports.
+   * The property to be used for sorting the records.
    */
   "sortBy"?: model.JavaDownloadRecordSortBy;
   /**
-   * The sort order for the reports.
+   * The sort order for the records.
    */
   "sortOrder"?: model.SortOrder;
   /**
    * The format of the report that is generated.
    */
   "format": model.JavaDownloadReportFormat;
+  /**
+   * Simple key-value pair that is applied without any predefined name, type, or scope. Exists for cross-compatibility only.
+   * Example: {@code {\"bar-key\": \"value\"}}. (See [Managing Tags and Tag Namespaces](https://docs.cloud.oracle.com/Content/Tagging/Concepts/understandingfreeformtags.htm).)
+   *
+   */
+  "freeformTags"?: { [key: string]: string };
+  /**
+   * Defined tags for this resource. Each key is predefined and scoped to a namespace.
+   * Example: {@code {\"foo-namespace\": {\"bar-key\": \"value\"}}}. (See [Understanding Free-form Tags](https://docs.cloud.oracle.com/Content/Tagging/Tasks/managingtagsandtagnamespaces.htm)).
+   *
+   */
+  "definedTags"?: { [key: string]: { [key: string]: any } };
 }
 
 export namespace CreateJavaDownloadReportDetails {